simple au double

+++ source/fonctions/euler.h
@@ -8,21 +8,21 @@ void euler(vec2 M[RES][RES])
     {
         for (size_t j = 0; j < RES; j++) // boucle for vecteur in mat_space
         {
-            float pos_x = M[i][j].x;
-            float pos_y = M[i][j].y;
-            float vx = v_0_x;
-            float vy = v_0_y;
-            for (float t = 0; t < t_f; t += dt)
+            double pos_x = M[i][j].x;
+            double pos_y = M[i][j].y;
+            double vx = v_0_x;
+            double vy = v_0_y;
+            for (double t = 0; t < t_f; t += dt)
             {
                 if (condition_capture(pos_x, pos_y)) // condition de capture toujours avant premier calcul pour éviter division par 0
                 {
                     break;
                 }
                 vec2 a = accel(pos_x, pos_y); // calcul du tableau accéleration
-                pos_x +=  vx * dt;
-                pos_y +=  vy * dt;
-                vx +=  a.x * dt;
-                vy +=  a.y * dt;
+                pos_x += vx * dt;
+                pos_y += vy * dt;
+                vx += a.x * dt;
+                vy += a.y * dt;
             }
             M[i][j].x = pos_x;
             M[i][j].y = pos_y;
